fixing openWeather mapper

+++ weather-you/data/src/main/java/com/rodrigmatrix/weatheryou/data/model/visualcrossing/CurrentConditionsResponse.kt
@@ -13,7 +13,7 @@ data class CurrentConditionsResponse(
     @SerialName("datetime")
     val datetime: String? = null,
     @SerialName("datetimeEpoch")
-    val datetimeEpoch: Int? = null,
+    val datetimeEpoch: Long? = null,
     @SerialName("dew")
     val dew: Double? = null,
     @SerialName("feelslike")
@@ -45,11 +45,11 @@ data class CurrentConditionsResponse(
     @SerialName("sunrise")
     val sunrise: String? = null,
     @SerialName("sunriseEpoch")
-    val sunriseEpoch: Int? = null,
+    val sunriseEpoch: Long? = null,
     @SerialName("sunset")
     val sunset: String? = null,
     @SerialName("sunsetEpoch")
-    val sunsetEpoch: Int? = null,
+    val sunsetEpoch: Long? = null,
     @SerialName("temp")
     val temp: Double? = null,
     @SerialName("uvindex")
